Harden Redis queues and Ubuntu firewall

+++ config/queue.php
@@ -69,7 +69,7 @@ return [
             'connection' => env('REDIS_QUEUE_CONNECTION', 'default'),
             'queue' => env('REDIS_QUEUE', 'default'),
             'retry_after' => (int) env('REDIS_QUEUE_RETRY_AFTER', 90),
-            'block_for' => env('REDIS_QUEUE_BLOCK_FOR', 5),
+            'block_for' => (int) env('REDIS_QUEUE_BLOCK_FOR', 5),
             'after_commit' => true,
         ],
